Simplify download API abstraction
