+++ src/cli/self_update/windows.rs
@@ -22,6 +22,7 @@ use super::common;
 use super::{InstallOpts, install_bins, report_error};
 use crate::cli::{download_tracker::DownloadTracker, markdown::md};
 use crate::dist::TargetTriple;
+use crate::download::download_file;
 use crate::process::{Process, terminalsource::ColorableTerminal};
 use crate::utils::{self, Notification};
 
@@ -276,7 +277,7 @@ pub(crate) async fn try_install_msvc(
     download_tracker.lock().unwrap().download_finished();
 
     info!("downloading Visual Studio installer");
-    utils::download_file(
+    download_file(
         &visual_studio_url,
         &visual_studio,
         None,
