+++ src/utils/mod.rs
@@ -11,9 +11,6 @@ use std::process::ExitStatus;
 use anyhow::{Context, Result, anyhow, bail};
 use retry::delay::{Fibonacci, jitter};
 use retry::{OperationResult, retry};
-use sha2::Sha256;
-#[cfg(any(feature = "reqwest-rustls-tls", feature = "reqwest-native-tls"))]
-use tracing::info;
 use url::Url;
 
 use crate::errors::*;
@@ -144,181 +141,6 @@ where
         notify_handler(Notification::NoCanonicalPath(path).into());
         PathBuf::from(path)
     })
-}
-
-pub async fn download_file(
-    url: &Url,
-    path: &Path,
-    hasher: Option<&mut Sha256>,
-    notify_handler: &dyn Fn(Notification<'_>),
-    process: &Process,
-) -> Result<()> {
-    download_file_with_resume(url, path, hasher, false, &notify_handler, process).await
-}
-
-pub(crate) async fn download_file_with_resume(
-    url: &Url,
-    path: &Path,
-    hasher: Option<&mut Sha256>,
-    resume_from_partial: bool,
-    notify_handler: &dyn Fn(Notification<'_>),
-    process: &Process,
-) -> Result<()> {
-    use crate::download::DownloadError as DEK;
-    match download_file_(
-        url,
-        path,
-        hasher,
-        resume_from_partial,
-        notify_handler,
-        process,
-    )
-    .await
-    {
-        Ok(_) => Ok(()),
-        Err(e) => {
-            if e.downcast_ref::<std::io::Error>().is_some() {
-                return Err(e);
-            }
-            let is_client_error = match e.downcast_ref::<DEK>() {
-                // Specifically treat the bad partial range error as not our
-                // fault in case it was something odd which happened.
-                Some(DEK::HttpStatus(416)) => false,
-                Some(DEK::HttpStatus(400..=499)) | Some(DEK::FileNotFound) => true,
-                _ => false,
-            };
-            Err(e).with_context(|| {
-                if is_client_error {
-                    RustupError::DownloadNotExists {
-                        url: url.clone(),
-                        path: path.to_path_buf(),
-                    }
-                } else {
-                    RustupError::DownloadingFile {
-                        url: url.clone(),
-                        path: path.to_path_buf(),
-                    }
-                }
-            })
-        }
-    }
-}
-
-async fn download_file_(
-    url: &Url,
-    path: &Path,
-    hasher: Option<&mut Sha256>,
-    resume_from_partial: bool,
-    notify_handler: &dyn Fn(Notification<'_>),
-    process: &Process,
-) -> Result<()> {
-    #[cfg(any(feature = "reqwest-rustls-tls", feature = "reqwest-native-tls"))]
-    use crate::download::{TlsBackend, Backend, Event};
-    use sha2::Digest;
-    use std::cell::RefCell;
-
-    notify_handler(Notification::DownloadingFile(url, path));
-
-    let hasher = RefCell::new(hasher);
-
-    // This callback will write the download to disk and optionally
-    // hash the contents, then forward the notification up the stack
-    let callback: &dyn Fn(Event<'_>) -> anyhow::Result<()> = &|msg| {
-        if let Event::DownloadDataReceived(data) = msg {
-            if let Some(h) = hasher.borrow_mut().as_mut() {
-                h.update(data);
-            }
-        }
-
-        match msg {
-            Event::DownloadContentLengthReceived(len) => {
-                notify_handler(Notification::DownloadContentLengthReceived(len));
-            }
-            Event::DownloadDataReceived(data) => {
-                notify_handler(Notification::DownloadDataReceived(data));
-            }
-            Event::ResumingPartialDownload => {
-                notify_handler(Notification::ResumingPartialDownload);
-            }
-        }
-
-        Ok(())
-    };
-
-    // Download the file
-
-    // Keep the curl env var around for a bit
-    let use_curl_backend = process.var_os("RUSTUP_USE_CURL").map(|it| it != "0");
-    let use_rustls = process.var_os("RUSTUP_USE_RUSTLS").map(|it| it != "0");
-
-    let backend = match (use_curl_backend, use_rustls) {
-        // If environment specifies a backend that's unavailable, error out
-        #[cfg(not(feature = "reqwest-rustls-tls"))]
-        (_, Some(true)) => {
-            return Err(anyhow!(
-                "RUSTUP_USE_RUSTLS is set, but this rustup distribution was not built with the reqwest-rustls-tls feature"
-            ));
-        }
-        #[cfg(not(feature = "reqwest-native-tls"))]
-        (_, Some(false)) => {
-            return Err(anyhow!(
-                "RUSTUP_USE_RUSTLS is set to false, but this rustup distribution was not built with the reqwest-native-tls feature"
-            ));
-        }
-        #[cfg(not(feature = "curl-backend"))]
-        (Some(true), _) => {
-            return Err(anyhow!(
-                "RUSTUP_USE_CURL is set, but this rustup distribution was not built with the curl-backend feature"
-            ));
-        }
-
-        // Positive selections, from least preferred to most preferred
-        #[cfg(feature = "curl-backend")]
-        (Some(true), None) => Backend::Curl,
-        #[cfg(feature = "reqwest-native-tls")]
-        (_, Some(false)) => {
-            if use_curl_backend == Some(true) {
-                info!(
-                    "RUSTUP_USE_CURL is set and RUSTUP_USE_RUSTLS is set to off, using reqwest with native-tls"
-                );
-            }
-            Backend::Reqwest(TlsBackend::NativeTls)
-        }
-        #[cfg(feature = "reqwest-rustls-tls")]
-        _ => {
-            if use_curl_backend == Some(true) {
-                info!(
-                    "both RUSTUP_USE_CURL and RUSTUP_USE_RUSTLS are set, using reqwest with rustls"
-                );
-            }
-            Backend::Reqwest(TlsBackend::Rustls)
-        }
-
-        // Falling back if only one backend is available
-        #[cfg(all(not(feature = "reqwest-rustls-tls"), feature = "reqwest-native-tls"))]
-        _ => Backend::Reqwest(TlsBackend::NativeTls),
-        #[cfg(all(
-            not(feature = "reqwest-rustls-tls"),
-            not(feature = "reqwest-native-tls"),
-            feature = "curl-backend"
-        ))]
-        _ => Backend::Curl,
-    };
-
-    notify_handler(match backend {
-        #[cfg(feature = "curl-backend")]
-        Backend::Curl => Notification::UsingCurl,
-        #[cfg(any(feature = "reqwest-rustls-tls", feature = "reqwest-native-tls"))]
-        Backend::Reqwest(_) => Notification::UsingReqwest,
-    });
-
-    let res = backend
-        .download_to_path(url, path, resume_from_partial, Some(callback))
-        .await;
-
-    notify_handler(Notification::DownloadFinished);
-
-    res
 }
 
 pub(crate) fn parse_url(url: &str) -> Result<Url> {
